Move utils.js from lib/conversations/ into lib/
