+++ lib/conversations/introduction.js
@@ -3,7 +3,7 @@ const {
   translate: t,
   conversations: {goto, utterances},
 } = require('borq');
-const {initConversation} = require('./utils.js');
+const {initConversation} = require('../utils.js');
 
 /**
  * Initializes and handles the conversations of the first bot
